+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- default `bit import` with no id to import objects only, unless `--write` flag is used
 - decrease verbosity of npm during bit test
 
 ## [0.12.5-dev.1] - 2018-02-05